aml: Add new and update existing modules
Add Android 14 modules. Update tz4 to tz5.

Some of these modules do not have any known short names yet. They have
seemingly not appeared with their own aml_ tags. So that we can build
them using the latest frc_ tag for now, we use their longer name as
the key.

One exception is Bluetooth, for which the frc_ tag is outdated and there
is no known appropriate tag, so we simply skip it.

Change-Id: I054fb903f787dd781d9a5024f27730cc7aca77fe